Fix deployment process
The new process requires changes to the launch script, as well as the
server config file. More specifically, it is no longer necessary for the
launch script to install requirements, build the front end, or purge the
cache. The config file will need to contain three new values:
auth_email, auth_key, and cloudflare_zone. These are used to purge the
cache after updating.